Restructure Makefile to permit installing all needed artifacts into a location other than the build directory
When you do a "make DESTDIR=<path> install", the following directory
structure will get created:

<destdir>/
  |
  +-- bin/
  | |
  | +-- rosie (script)
  +-- share/
       |
      rosie/
         |
         +--bin/
         |
         +--lib/
         |
         +--src/

This way if DESTDIR is specified as /usr/local, all of the artifacts will
end up in the right places.  So that's why the default location of
DESTDIR is changed to /usr/local from /usr/local/bin

The install is broken up into many separate targets, each of which
create the directories they need, which is sometimes redundant, but
if you run with -j8, for example, you will find that these targets are run
in parallel, and so there would be race conditions with directory creation
if these mkdir commands were confined to one target.

In addition to these changes:

I found that the variables names
HOME and TMP were confusing, so I renamed them to what I think are
more appropriate names: BUILD_ROOT and SUBMOD respectively.

The two variables: ROSIEBIN and EXECROSIE were confusing because
the target of the of the recipe was different from the file actually
being created, even though both variables really refer to the same
file on the disk.  So I got rid of EXECROSIE, and just use
ROSIEBIN everywhere now.  This doesn't seem to cause any problems.

Thanks for Kevin Zander here at IBM for working together with me
on these changes.