Issue #57
(1) Fixed a small memory leak where a string the length of a file name
was not freed upon creation of an engine using the "initialize" call.
If a client created and destroyed many engines, a few dozen bytes
would leak each time.

(2) Added more informational messages to the python client for
librosie, rtest.py, and to the "rough draft" python library, rosie.py.

(3) Calling this v0.99k-mod3