Found a reasonable compromise for the near future: (1) A macro invocation at top level generates an anonymous match (type="*"), as one would expect. (2) The argument to the "find" macro is, currently, ALWAYS treated as an anonymous expression.
This leads to the results shown below, in which there is redundant information in the match data,
because the anonymous interior match will always have one submatch called "word.any".  This is
the case that we optimize out when matching a reference at top level.  We detect that the top
level expression is a reference and use that instead of an anonymous exp containing only that ref.

It would be ideal if we could do the same here.  In the example below, however, we cannot give a
name to the macro argument without creating an assignment for it (a rule inside the grammar
generated by the macro).  And when we create an assignment inside a grammar, we run the risk of
capturing a free name (and can lead to a rule like 'a = a' and the error that the rule for 'a' can
match the empty string.

If we had gensym for the rule identifier, we would have, e.g. 'b43d78qq1 = a', and could verify
that the generated symbol does not capture any free variable.  But then we would see match output
with type="b43d78qq1" in every place that we want to see type="a".

Having an explicit "set a name for this capture" function available to macro implementors would be
one solution.  This has its own implementation challenges, but is ultimately the way to go.  This
would allow other cool things, like a macro that renames a capture:
E.g.      foo = rename(x, "y") y z
	  This would match 'x y z' but the matches that would normally have type="x"
	  would have type="y" instead.  And the second arg to rename could be any string.

$ bin/rosie match -o json 'findall:word.any' test/resolv.conf | tail -n 1 | rjsonpp
Loading from source
{"subs":
   [{"subs":
       [{"data": "nameserver",
         "end": 11.0,
         "pos": 1.0,
         "type": "word.any"}],
     "pos": 1.0,
     "data": "nameserver",
     "type": "*",
     "end": 11.0},
    {"subs":
       [{"data": "fde",
         "end": 15.0,
         "pos": 12.0,
         "type": "word.any"}],
     "pos": 12.0,
     "data": "fde",
     "type": "*",
     "end": 15.0},
    {"subs":
       [{"data": "dd",
         "end": 26.0,
         "pos": 24.0,
         "type": "word.any"}],
     "pos": 24.0,
     "data": "dd",
     "type": "*",
     "end": 26.0},
    {"subs":
       [{"data": "bd",
         "end": 31.0,
         "pos": 29.0,
         "type": "word.any"}],
     "pos": 29.0,
     "data": "bd",
     "type": "*",
     "end": 31.0}],
 "pos": 1.0,
 "data": "nameserver fde9:4789:96dd:03bd",
 "type": "*",
 "end": 31.0}
$